requirements.txt
paho-mqtt==1.6.1
protobuf==5.26.1
fastapi
uvicorn[standard]
